Refactor and move common classes
- Renamed a file to better reflect its purpose.
- Moved class definitions from one module to another for improved organization.
- Updated import statements in related modules to match the new structure.
